refactor(config): extract model layer from manager.py to _models.py
Extrai a camada de modelos de dados de configuração (FunctionCallingMode + 6
dataclasses) de deile/config/manager.py para o módulo irmão
deile/config/_models.py.

manager.py mantém a classe ConfigManager, o singleton get_config_manager() e
um shim de re-export que preserva a superfície pública existente sem quebrar
nenhum import.

O módulo _models.py é folha pura de stdlib (dataclasses, enum, typing),
sem dependências do DEILE — ciclo de import impossível por construção.

Refs #756
